api/requirements.txt
Django==1.6.2
Markdown==2.4
argparse==1.2.1
distribute==0.6.24
django-filter==0.7
djangorestframework==2.3.13
ipython==2.0.0
wsgiref==0.1.2
